* sdcc.sln, * src/stm8/stm8.vcxproj, * src/stm8/stm8.vcxproj.filters: added stm8 project * src/hc08/hc08.vcxproj, * src/z80/z80.vcxproj, * src/sdcc.vcxproj: use boost 1.50, the last working installer from boostpro * debugger/mcs51/symtab.h, * src/ds390/gen.c, * src/hc08/gen.c, * src/mcs51/gen.c, * src/pic14/gen.c, * src/pic16/gen.c, * src/pic16/glue.c, * src/SDCCast.c, * src/SDCCdwarf2.c, * src/SDCCglue.c, * src/SDCCsymt.h, * src/z80/gen.c: fixed signed/unsigned comparison warnings * src/stm8/gen.c: fixed errors for MSVC (C89 compatible only):   don't use snprintf but only SNPRINT,   only declarations at start of block,   don't use true/false but only TRUE/FALSE * src/stm8/main.c (stm8_defaultRules): fixed init error * src/stm8/ralloc2.cc (instruction_cost): fixed warning * src/SDCCglue.c (glue): removed target dependent code * support/regression/fwk/include/testfwk.h: removed ASSERT_FAILED,   added ASSERTFALSE * support/regression/tests/gcc-torture-execute-20050826-1.c: use ASSERFALSE,   enabled test for mcs51 * support/regression/tests/gcc-torture-execute-920810-1.c: use ASSERTFALSE
git-svn-id: http://svn.code.sf.net/p/sdcc/code/trunk@8708 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ sdcc/src/stm8/ralloc2.cc
@@ -305,7 +305,7 @@ static float instruction_cost(const assignment &a, unsigned short int i, const G
     case SWAP:
       assign_operands_for_cost(a, i, G, I);
       set_surviving_regs(a, i, G, I);
-      c = drySTM8iCode(ic);
+      c = (float)drySTM8iCode(ic);
       unset_surviving_regs(i, G);
 #if 0
       std::cout << "Got cost " << c << "\n";
